refactor(SchedulerContainer): move it in a separate file
