+++ src/engine/src/core/Core.cpp
@@ -1,4 +1,7 @@
 #include "Core.hpp"
+#include "FixedTimeUpdate.hpp"
+#include "RelativeTimeUpdate.hpp"
+#include "Startup.hpp"
 #include "Entity.hpp"
 
 ES::Engine::Core::Core() : _registry(nullptr)
